backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1u
ultralytics>=8.1.0
numpy>=1.26.0
pillow>=10.2.0
opencv-python-headless>=4.9.0
geopy>=2.4.0
psycopg2